+++ backend/user-center/src/main/java/vn/winter/usercenter/otp/OtpRepository.java
@@ -1,0 +1,6 @@
+package vn.winter.usercenter.otp;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface OtpRepository extends JpaRepository<OTP, Long> {
+}