+++ backend/user-center/src/main/java/vn/winter/usercenter/user/UserController.java
@@ -4,9 +4,13 @@ import jakarta.validation.Valid;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
+import vn.winter.usercenter.otp.OTP;
 import vn.winter.usercenter.user.dto.UserSignInDto;
 import vn.winter.usercenter.user.dto.UserSignUpDto;
+import vn.winter.usercenter.util.OtpType;
+import vn.winter.usercenter.util.Util;
 
+import java.time.LocalDateTime;
 import java.util.List;
 
 @RestController()
@@ -33,5 +37,12 @@ public class UserController {
     public ResponseEntity<Object> signIn(@Valid @RequestBody UserSignInDto userSignInDto) {
         return this.userService.signIn(userSignInDto);
     }
+
+    // Active user
+    // Re-send OTP
+    // Forget password
+    // Change password
+    // Change information
+    // ...
 
 }