+++ backend/user-center/src/main/java/vn/winter/usercenter/user/dto/UserDto.java
@@ -6,6 +6,8 @@ import lombok.NoArgsConstructor;
 import lombok.Setter;
 
 import java.io.Serializable;
+import java.time.LocalDate;
+import java.time.LocalDateTime;
 import java.util.Date;
 
 @Getter
@@ -17,9 +19,11 @@ public class UserDto implements Serializable {
     private String fullName;
     private String email;
     private String phone;
+    private LocalDate birthday;
     private String avatarUrl;
-    private Date createdAt;
-    private Date updatedAt;
+    private LocalDateTime createdAt;
+    private LocalDateTime updatedAt;
     private boolean isDeleted;
     private boolean isBlocked;
+    private boolean isActive;
 }