+++ backend/user-center/src/main/java/vn/winter/usercenter/util/OtpType.java
@@ -1,0 +1,6 @@
+package vn.winter.usercenter.util;
+
+public enum OtpType {
+    REGISTER,
+    FORGET_PASSWORD
+}