+++ database/user-center/01.add-field-for-user-table.sql
@@ -1,0 +1,5 @@
+-- IsActive field
+alter table users add column "is_active" boolean default false;
+
+-- Birthday field
+alter table users add column "birthday" date not null;