+++ database/CHANGELOG.md
@@ -1,5 +1,15 @@
 # Changelog
 
+## 2023 - 12 - 24
+- Add "is_active" field for "users" table.
+- Add "birthday" field for "users" table.
+
+## 2023 - 12 - 19
+- Insert default value (Free storage plan).
+
+## 2023 - 12 - 03
+- Initialize the database (User center + File manager).
+
 ## YYYY - MM - DD
 ### Content